Separa em mais pacotes

+++ crypto/rand.go
@@ -26,6 +26,6 @@ func randomPrime() *big.Int {
 }
 
 
-func number(n... int64) *big.Int {
-	return big.NewInt(append(n, 0)[0])
+func number(n... int) *big.Int {
+	return big.NewInt(int64(append(n, 0)[0]))
 }
